# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Private API - Quiz Management System
  description: |
    # Quiz Management System API

    A comprehensive RESTful API for managing educational quizzes, assessments, and submissions.

    ## Architecture

    This API follows **Domain-Driven Design (DDD)** principles with proper bounded contexts:

    - **Authentication**: User registration, login, session management
    - **Subject**: Educational subject management (Math, Physics, etc.)
    - **Grade**: Grade level management (Grade 10, Grade 11, etc.)
    - **Module**: Quiz/Exam content management with questions and choices
    - **Submission**: Public quiz-taking and answer submission

    ## Authentication

    Protected endpoints require a JWT Bearer token in the Authorization header:

    ```
    Authorization: Bearer <access_token>
    ```

  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.private.arvinpaundra.dev
    description: Production server
security:
  - BearerAuth: []
...
paths:
...
  /v1/submissions:
    get:
      tags:
        - Submissions
      summary: Get all submissions grouped by modules (Admin)
      description: |
        Retrieves all submitted submissions grouped by modules with their associated grade and subject.
        Returns module details, total submissions count, and list of students who completed each module.
        Only returns submissions with status "submitted".
      operationId: getAllSubmissions
      parameters:
        - $ref: '#/components/parameters/Keyword'
      responses:
        '200':
          description: Submissions retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ModuleSubmissionGroup'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==========================================
  # Quiz Submission (Public)
  # ==========================================
  /v1/modules/{module_slug}/submissions:
    post:
      tags:
        - Submissions
      summary: Start a new quiz submission
      description: Creates a new submission for a student to take the quiz
      operationId: startSubmission
      security: []
      parameters:
        - $ref: '#/components/parameters/ModuleSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartSubmissionRequest'
      responses:
        '201':
          description: Submission started successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/StartSubmissionResponse'
              example:
                meta:
                  code: 201
                  message: submission started successfully
                data:
                  code: 'A1B2C3D4E5F6G7H8'
                  status: 'in_progress'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/modules/{module_slug}/submissions/{submission_code}/answers:
    post:
      tags:
        - Submissions
      summary: Submit an answer to a question
      description: Submits a student's answer for a specific question
      operationId: submitAnswer
      security: []
      parameters:
        - $ref: '#/components/parameters/ModuleSlug'
        - $ref: '#/components/parameters/SubmissionCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAnswerRequest'
      responses:
        '200':
          description: Answer submitted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SubmitAnswerResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/modules/{module_slug}/submissions/{submission_code}/finalize:
    patch:
      tags:
        - Submissions
      summary: Finalize a quiz submission
      description: Finalizes a quiz submission and calculates the final score
      operationId: finalizeSubmission
      security: []
      parameters:
        - $ref: '#/components/parameters/ModuleSlug'
        - $ref: '#/components/parameters/SubmissionCode'
      responses:
        '200':
          description: Submission finalized successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FinalizeSubmissionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  parameters:
    ModuleSlug:
      name: module_slug
      in: path
      required: true
      description: Slug of the module
      schema:
        type: string
        example: 'introduction-to-algebra'
    SubmissionCode:
      name: submission_code
      in: path
      required: true
      description: Unique submission code
      schema:
        type: string
        example: 'A1B2C3D4E5F6G7H8'
...
    Keyword:
      name: keyword
      in: query
      description: Search keyword
      schema:
        type: string
        example: 'math'
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            meta:
              code: 400
              message: invalid request body
            data: null
            errors:
              email: email is required
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            meta:
              code: 401
              message: unauthorized
            data: null
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            meta:
              code: 404
              message: resource not found
            data: null
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            meta:
              code: 409
              message: resource already exists
            data: null
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            meta:
              code: 500
              message: internal server error
            data: null
  schemas:
...
    Meta:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: operation successful
      required:
        - code
        - message
...
    SuccessResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: object
          nullable: true
          description: Response data varies by endpoint
      required:
        - meta
        - data
...
    ErrorResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: object
          nullable: true
      required:
        - meta
        - data
    ValidationErrorResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: object
          nullable: true
        errors:
          type: object
          additionalProperties:
            type: string
          example:
            email: email is required
            password: password must be at least 8 characters
      required:
        - meta
        - data
...
    StartSubmissionRequest:
      type: object
      properties:
        student_name:
          type: string
          minLength: 3
          maxLength: 100
          example: 'John Doe'
      required:
        - student_name
    StartSubmissionResponse:
      type: object
      properties:
        code:
          type: string
          example: 'A1B2C3D4E5F6G7H8'
        status:
          type: string
          enum: [in_progress, submitted, canceled]
          example: 'in_progress'
      required:
        - code
        - status
    SubmitAnswerRequest:
      type: object
      properties:
        question_slug:
          type: string
          example: 'question-1'
        choice_slug:
          type: string
          example: 'choice-a'
      required:
        - question_slug
        - choice_slug
    SubmitAnswerResponse:
      type: object
      properties:
        is_correct:
          type: boolean
          example: true
        answer:
          type: string
          example: 'Paris'
        next_question_slug:
          type: string
          nullable: true
          example: 'question-2'
      required:
        - is_correct
        - answer
    FinalizeSubmissionResponse:
      type: object
      properties:
        student_name:
          type: string
          example: 'John Doe'
        score:
          type: integer
          minimum: 0
          example: 8
        total:
          type: integer
          minimum: 0
          example: 10
        status:
          type: string
          enum: [submitted]
          example: 'submitted'
      required:
        - student_name
        - score
        - total
        - status
    ModuleSubmissionGroup:
      type: object
      properties:
        module:
          $ref: '#/components/schemas/ModuleWithRelations'
        total_submissions:
          type: integer
          description: Total number of submitted submissions for this module
          example: 15
        submissions:
          type: array
          description: List of students who submitted the module
          items:
            $ref: '#/components/schemas/SubmissionSummary'
      required:
        - module
        - total_submissions
        - submissions
    ModuleWithRelations:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        title:
          type: string
          example: 'Introduction to Algebra'
        slug:
          type: string
          example: 'introduction-to-algebra'
        grade:
          $ref: '#/components/schemas/GradeInfo'
        subject:
          $ref: '#/components/schemas/SubjectInfo'
      required:
        - id
        - title
        - slug
        - grade
        - subject
    GradeInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        name:
          type: string
          example: 'Grade 10'
      required:
        - id
        - name
    SubjectInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        name:
          type: string
          example: 'Mathematics'
      required:
        - id
        - name
    SubmissionSummary:
      type: object
      properties:
        student_name:
          type: string
          description: Name of the student who submitted
          example: 'John Doe'
        score:
          type: number
          format: float
          description: Score as a decimal between 0.0 and 1.0 (correct answers / total questions)
          minimum: 0.0
          maximum: 1.0
          example: 0.85
        submitted_at:
          type: string
          description: Submission timestamp in format YYYY-MM-DD HH:MM:SS
          example: '2026-01-02 10:30:45'
      required:
        - student_name
        - score
        - submitted_at